+++ SAMuel/Modules/KofaxModule.vb
@@ -9,6 +9,7 @@ Module KofaxModule
     ''' List of per document tiff files stored in another list
     ''' </param>
     Sub CreateXML(ByRef docList As List(Of List(Of String)), batchName As String)
+        '' Deprecated overload. Removing soon.
         Dim sPath As String
         Dim sFile As String
         Dim sList As List(Of String)
@@ -60,6 +61,38 @@ Module KofaxModule
         Dim sAccountNumber As String
         Dim _IsCustomerNumber As Boolean
         Dim batchClass As String
+        Dim accountDictionary As Dictionary(Of String, List(Of String)) = New Dictionary(Of String, List(Of String))
+
+
+        'Write each tiff file as a document
+        For Each value In docList
+            _IsCustomerNumber = False
+
+            sAccountNumber = RegexAcc(value, "\d{5}-\d{5}")
+            If sAccountNumber = "X" Then
+                sAccountNumber = RegexAcc(value, "\d{10}")
+                If sAccountNumber = "X" Then
+                    sAccountNumber = RegexAcc(value, "\d{9}")
+                    If sAccountNumber = "X" Then
+                        sAccountNumber = "UNKNOWN"
+                    End If
+                End If
+            End If
+
+            'Remove Hyphen from account number
+            sAccountNumber = sAccountNumber.Replace("-", "")
+
+            If accountDictionary.ContainsKey(sAccountNumber) Then
+                'Add the file to the existing account number list
+                accountDictionary.Item(sAccountNumber).Add(value)
+            Else
+                'Add the file to a list and add the account with the list to the dictionary.
+                Dim fileList As List(Of String) = New List(Of String)
+                fileList.Add(value)
+                accountDictionary.Add(sAccountNumber, fileList)
+            End If
+
+        Next
 
         sPath = "C:\ACXMLAID\" 'Default Kofax Import Connector watch folder
         sFile = sPath & batchName & "-import.xml"
@@ -80,42 +113,51 @@ Module KofaxModule
         s1.Write(vbTab & vbTab & "<Batch BatchClassName=""" & batchClass & """ EnableSingleDocProcessing = ""0"" Name=""" & batchName & """>")
         s1.Write(vbTab & vbTab & vbTab & "<Documents>")
 
+        'Iterate through the dictonary
+        Dim acc As KeyValuePair(Of String, List(Of String))
+        For Each acc In accountDictionary
+            Dim fileEntry As String
+            'Get the account number
+            Dim account As String = acc.Key
 
-
-        'Write each tiff file as a document
-        For Each value In docList
-
-            _IsCustomerNumber = False
-
-            sAccountNumber = RegexAcc(value, "\d{5}-\d{5}")
-            If sAccountNumber = "X" Then
-                sAccountNumber = RegexAcc(value, "\d{10}")
-                If sAccountNumber = "X" Then
-                    sAccountNumber = RegexAcc(value, "\d{9}")
-                    If sAccountNumber = "X" Then
-                        sAccountNumber = ""
-                    Else
-                        _IsCustomerNumber = True
-                    End If
-                End If
+            'Determine the account type. Customer/Bill Account
+            If account.Length = 10 Then
+                'Is a bill account number
+                _IsCustomerNumber = False
+            ElseIf account.Length = 9 Then
+                'Is  a customer number
+                _IsCustomerNumber = True
+            ElseIf account = "UNKNOWN" Then
+                'There was no account number
+                'Doesn't matter what it's inserted into.
+                _IsCustomerNumber = False
+            Else
+                'This case shouldn't be reached but..
+                _IsCustomerNumber = False
+                LogAction(99, String.Format("Kofax it account number type else case reached. Account: {0} File: {1}", account, acc.Value.Item(0)))
             End If
 
-            'Remove Hyphen from account number
-            sAccountNumber = sAccountNumber.Replace("-", "")
-
+            'Write the XML File
+            'TODO Use DOM
             s1.Write(vbTab & vbTab & vbTab & vbTab & "<Document FormType = ""eCorrespondence Fax"">")
             s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexFields>")
             If (_IsCustomerNumber) Then
-                s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexField Name=""Customer Account"" Value = """ & sAccountNumber & """ />")
+                s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexField Name=""Customer Account"" Value = """ & account & """ />")
             Else
-                s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexField Name=""Bill Account"" Value = """ & sAccountNumber & """ />")
+                s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexField Name=""Bill Account"" Value = """ & account & """ />")
             End If
             s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexField Name=""Correspondence Type"" Value = """ & corrType & """ />")
             s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexField Name=""Correspondence Source"" Value = """ & corrSource & """ />")
             s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<IndexField Name=""Comments"" Value = """ & strComments & """ />")
             s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & "</IndexFields>")
             s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & "<Pages>")
-            s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<Page ImportFileName=""" & value & """/>")
+
+            'Add each file associated to the account, as a page in kofax. 
+            ''*NOTE* A page in kofax can be one or more actual pages.
+            For Each fileEntry In acc.Value
+                s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & vbTab & "<Page ImportFileName=""" & fileEntry & """/>")
+            Next
+
             s1.Write(vbTab & vbTab & vbTab & vbTab & vbTab & "</Pages>")
             s1.Write(vbTab & vbTab & vbTab & vbTab & "</Document>")
         Next