+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 2014-08-19
+
+Features:
+
+  - Kofax It: Now groups 'pages' or files with the same account number as a single document.
+  - Kofax It: Added several correspondance types.
+  - Add Contact: Now functions correctly using AHK script with commandline arguments.
+
+Code Improvements:
+
+  - Kofax It: Added dictonaries to handle each account and associated files to be uploaded together.
+  - Kofax It: Now logs progress when emails canceled during processing.
+  - SAM_Email Class: Added .txt as known format.
+
+UI Improvements:
+
+  - Kofax It: Added status text when completed.
+  - General: Added welcome.
+  - Convert Files: Added status text when converting.
+  - Add Contact: Button disable while script is running.
+
 ## 2014-08-17
 
 Features: